instead of router using service
