+++ src/modules/auth/auth.controller.js
@@ -1,39 +1,28 @@
-import bcrypt from "bcrypt";
-import prisma from "../../database/prisma.js";
+import { Router } from "express";
+import AuthServices from "./auth.services.js";
 
-const login = async (email, password) => {
-    const user = await prisma.user.findFirst({
-        where: { email }
-    });
-    if (!user) {
-        throw new Error("User not found");
+const router = Router();
+
+router.post("/login", async (req, res, next) => {
+    try {
+        const { email, password } = req.body;
+        if (!email || !password) {
+            throw new Error("Email and password are required");
+        }
+        const result = await AuthServices.login(email, password);
+        res.json({ data: result, msg: "Logged in Successfully" });
+    } catch(e) {
+        res.status(400).json({ msg: e.message });
     }
-    const isMatch = await bcrypt.compare(password, user.password);
-    if (!isMatch) {
-        throw new Error("Incorrect password");
+});
+
+router.post("/register", async (req, res, next) => {
+    try {
+        const result = await AuthServices.register(req.body);
+        res.json({ data: result, msg: "User Successfully Created" });
+    } catch(e) {
+        res.status(400).json({ msg: e.message });
     }
-    return user;
-};
+});
 
-const register = async (payload) => {
-      const { email, password } = payload;
-      const user = await prisma.user.findFirst({
-        where: { email },
-      });
-      if (user) {
-        throw new Error("User already exists");
-      }
-      const hashedPassword = await bcrypt.hash(password, 10);
-      const newUser = await prisma.user.create({
-        data: {
-          email,
-          password: hashedPassword,
-        },
-      });
-      return newUser;
-};
-
-export default {
-    login,
-    register,
-};
+export default router;
