--- src/modules/users/user.routes.js
@@ -1,42 +1,0 @@
-import express from "express";
-import Controller from "./user.controller.js";
-
-const router = express.Router();
-
-router.get("/", async (req, res, next) => {
-    try {
-        const result = await Controller.getAll();
-        res.json({ data: result, msg: "success" });
-    } catch (e) {
-        res.status(400).json({ msg: e.message });
-    }
-})
-
-router.get("/:id", async (req, res, next) => {
-  try {
-    const result = await Controller.getById(req.params.id);
-    res.json({ data: result, msg: "success" });
-  } catch (e) {
-    res.status(400).json({ msg: e.message });
-  }
-});
-
-router.post("/", async (req, res, next) => {
-  try {
-    const result = await Controller.create(req.body);
-    res.json({ data: result, msg: "success" });
-  } catch (e) {
-    res.status(400).json({ msg: e.message });
-  }
-});
-
-router.put("/:id", async (req, res, next) => {
-  try {
-    const result = await Controller.updateById(req.params.id, req.body);
-    res.json({ data: result, msg: "success" });
-  } catch (e) {
-    res.status(400).json({ msg: e.message });
-  }
-});
-
-export default router;
