+++ src/routes/routes.api.js
@@ -1,6 +1,6 @@
 import express from "express";
-import userRouter from "../modules/users/user.routes.js";
-import authRouter from "../modules/auth/auth.routes.js"
+import userController from "../modules/users/user.controller.js";
+import authController from "../modules/auth/auth.controller.js";
 
 const router = express.Router();
 
@@ -8,8 +8,8 @@ router.get("/", (req, res, next) => {
   res.json({ data: "", msg: "API Router is working" });
 });
 
-router.use("/users", userRouter);
-router.use("/auth", authRouter);
+router.use("/users", userController);
+router.use("/auth", authController);
 
 router.all("*", (req, res, next) => {
   res.json({ data: "", msg: "Route not found..." });
